tctl resources: convert users
Convert users to the new tctl resource handler format that
was introduced in #59518.
